requirements.txt
accelerate==0.21.0
aiohttp==3.8.1
aiosignal==1.2.0
altair==5.0.1
antlr4-python3-runtime==4.9.3
anyio==3.6.2
appdirs==1.4.4
arrow==1.2.3
asttokens==2.0.5
async-timeout==4.0.2
atomicwrites==1.1.5
attrs==23.1.0
backcall==0.2.0
backports.zoneinfo==0.2.1
bcrypt==3.1.7
beautifulsoup4==4.12.2
bitsandbytes==0.38.1
bleach==6.0.0
blessed==1.20.0
blessings==1.7
blinker==1.4
blis==0.7.10
braceexpand==0.1.7
cachetools==5.3.1
catalogue==2.0.9
certifi==2019.11.28
cfgv==3.3.1
chardet==3.0.4
charset-normalizer==2.1.1
Click==7.0
cmake==3.26.3
colorama==0.4.3
confection==0.1.0
contexttimer==0.3.3
contourpy==1.1.0
croniter==1.3.14
cryptography==2.8
cupshelpers==1.0
cycler==0.11.0
cymem==2.0.7
datasets==2.11.0
dateutils==0.6.12
dbus-python==1.2.16
debugpy==1.6.0
decorator==5.1.1
decord==0.6.0
deepdiff==6.3.0
deepspeed==0.9.0
defer==1.0.6
diffusers==0.16.0
dill==0.3.6
distlib==0.3.7
distro==1.4.0
distro-info===0.23ubuntu1
docker-pycreds==0.4.0
docstring-parser==0.15
duplicity==0.8.12.0
editdistance==0.6.2
einops==0.6.1
entrypoints==0.3
exceptiongroup==1.1.1
executing==0.8.3
fabric==3.1.0
fairscale==0.4.4
fastapi==0.93.0
fasteners==0.14.1
fastmri==0.2.0
filelock==3.11.0
fonttools==4.33.3
frozenlist==1.3.0
fsspec==2023.4.0
ftfy==6.1.1
future==0.18.2
gitdb==4.0.10
GitPython==3.1.31
google-auth-oauthlib==0.4.6
gpustat==0.6.0
h11==0.14.0
h5py==3.7.0
hjson==3.1.0
httplib2==0.14.0
huggingface-hub==0.14.1
identify==2.5.26
idna==2.8
imageio==2.31.1
importlib-metadata==1.5.0
importlib-resources==5.12.0
iniconfig==2.0.0
inquirer==3.1.3
invoke==2.1.2
iopath==0.1.10
ipykernel==6.15.0
ipython==8.4.0
itsdangerous==2.1.2
jedi==0.18.1
Jinja2==3.1.2
joblib==1.3.2
jsonargparse==4.20.1
jsonschema==4.18.4
jsonschema-specifications==2023.7.1
jupyter-client==7.3.4
jupyter-core==4.10.0
kaggle==1.5.16
keyring==18.0.1
kiwisolver==1.4.4
langcodes==3.3.0
language-selector==0.1
launchpadlib==1.10.13
lazr.restfulclient==0.14.2
lazr.uri==1.0.3
lazy-loader==0.3
lightning==2.0.1.post0
lightning-cloud==0.5.33
lightning-utilities==0.8.0
lit==16.0.1
lockfile==0.12.2
louis==3.12.0
macaroonbakery==1.3.1
Mako==1.1.0
markdown-it-py==2.2.0
MarkupSafe==2.1.2
matplotlib==3.7.2
matplotlib-inline==0.1.3
mdurl==0.1.2
monotonic==1.5
more-itertools==4.2.0
mpmath==1.3.0
multidict==6.0.2
multiprocess==0.70.14
murmurhash==1.0.9
nest-asyncio==1.5.5
netifaces==0.10.4
networkx==3.1
ninja==1.11.1
nodeenv==1.8.0
numpy==1.23.3
nvidia-cublas-cu11==11.10.3.66
nvidia-cuda-cupti-cu11==11.7.101
nvidia-cuda-nvrtc-cu11==11.7.99
nvidia-cuda-runtime-cu11==11.7.99
nvidia-cudnn-cu11==8.5.0.96
nvidia-cufft-cu11==10.9.0.58
nvidia-curand-cu11==10.2.10.91
nvidia-cusolver-cu11==11.4.0.1
nvidia-cusparse-cu11==11.7.4.91
nvidia-ml-py3==7.352.0
nvidia-nccl-cu11==2.14.3
nvidia-nvtx-cu11==11.7.91
oauthlib==3.1.0
olefile==0.46
omegaconf==2.3.0
openai==0.27.8
opencv-python-headless==4.5.5.64
opendatasets==0.1.22
ordered-set==4.1.0
packaging==21.3
pandas==2.0.0
paramiko==2.6.0
parso==0.8.3
pathtools==0.1.2
pathy==0.10.2
peft==0.4.0
pexpect==4.6.0
pickleshare==0.7.5
Pillow==9.5.0
pkgutil-resolve-name==1.3.10
platformdirs==3.9.1
plotly==5.15.0
pluggy==0.13.0
portalocker==2.7.0
pre-commit==3.3.3
preshed==3.0.8
prompt-toolkit==3.0.29
protobuf==4.23.2
psutil==5.9.1
pure-eval==0.2.2
py==1.8.1
py-cpuinfo==9.0.0
pyarrow==11.0.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycairo==1.16.2
pycocoevalcap==1.2
pycocotools==2.0.6
pycups==1.9.73
pydantic==1.10.7
pydeck==0.8.0
Pygments==2.15.0
PyGObject==3.36.0
PyJWT==1.7.1
pymacaroons==0.13.0
Pympler==1.0.1
PyNaCl==1.3.0
pyparsing==3.0.9
pypdf2==3.0.1
pyRFC3339==1.1
pytest==7.2.1
python-apt==2.0.1+ubuntu0.20.4.1
python-dateutil==2.8.2
python-debian===0.1.36ubuntu1
python-dotenv==1.0.0
python-editor==1.0.4
python-magic==0.4.27
python-multipart==0.0.6
python-slugify==8.0.1
pytorch-lightning==2.0.1.post0
pytz==2022.1
pytz-deprecation-shim==0.1.0.post0
PyWavelets==1.3.0
pyxdg==0.26
PyYAML==6.0
pyzmq==23.2.0
rank-bm25==0.2.2
readchar==4.0.5
referencing==0.30.0
regex==2023.5.5
reportlab==3.5.34
requests==2.31.0
requests-oauthlib==1.3.1
requests-unixsocket==0.2.0
responses==0.18.0
rich==13.3.4
rpds-py==0.9.2
rsa==4.8
runstats==2.0.0
safetensors==0.3.2
-e git+https://github.com/salesforce/LAVIS.git@91c8e6863b4b02d7d75167e7d18037ef3a96c54b#egg=salesforce_lavis
scikit-image==0.21.0
scikit-learn==1.3.0
scipy==1.9.1
screen-resolution-extra==0.0.0
SecretStorage==2.3.1
sentencepiece==0.1.98
sentry-sdk==1.25.0
setproctitle==1.3.2
simplejson==3.16.0
six==1.14.0
sklearn==0.0.post7
smart-open==6.3.0
smmap==5.0.0
sniffio==1.3.0
soupsieve==2.4
spacy==3.6.0
spacy-legacy==3.0.12
spacy-loggers==1.0.4
srsly==2.4.7
ssh-import-id==5.10
stack-data==0.3.0
starlette==0.25.0
starsessions==1.3.0
streamlit==1.25.0
sympy==1.11.1
systemd-python==234
tenacity==8.2.2
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
text-unidecode==1.3
thinc==8.1.10
threadpoolctl==3.2.0
tifffile==2023.7.10
tiktoken==0.4.0
timm==0.4.12
tokenizers==0.13.3
toml==0.10.2
tomli==2.0.1
toolz==0.12.0
torch==2.0.1
torchmetrics==0.11.4
torchvision==0.15.2
tornado==6.1
tqdm==4.65.0
traitlets==5.3.0
transformers==4.29.2
triton==2.0.0
typer==0.9.0
typeshed-client==2.2.0
typing-extensions==4.3.0
tzdata==2023.3
tzlocal==4.3.1
ubuntu-advantage-tools==8001
ubuntu-drivers-common==0.0.0
ufw==0.36
unattended-upgrades==0.1
urllib3==2.0.2
usb-creator==0.3.7
uvicorn==0.21.1
validators==0.20.0
virtualenv==20.24.2
wadllib==1.3.3
wandb==0.15.3
wasabi==1.1.2
watchdog==3.0.0
wcwidth==0.2.5
webdataset==0.2.48
webencodings==0.5.1
websocket-client==1.5.1
websockets==11.0.1
xkit==0.0.0
xxhash==3.2.0
yarl==1.7.2
zipp==3.15.0
zstandard==0.21.0
